Refactor: component imports to use lowercase folder names and changed folder organization for buttons and animations
