+++ src/app/docs/animations/hovers/vibrate/page.mdx
@@ -1,7 +1,7 @@
 import Link from 'next/link'
 import { RequiresTailwind } from '@/components/Requires'
-import HoverAnimations from '@/components/HoverAnimationCards'
-import { HoverVibrateButton } from '@/components/PreviewCode/HoverAnimationButtons'
+import HoverAnimations from '@/components/cards/HoverAnimationCards'
+import { HoverVibrateButton } from '@/showcase/animations/hover/HoverAnimationButtons'
 import { ChevronLeft } from 'lucide-react'
 import PreviewCodeTabs from '@/components/CodePreview/page'
