+++ src/app/docs/components/button/heartbeat-button/page.mdx
@@ -1,4 +1,4 @@
-import HeartbeatButton from '@/ui/animation/HeartbeatButton'
+import HeartbeatButton from '@//showcase/components/button/HeartbeatButton'
 import { RequiresTailwind } from '@/components/Requires'
 import Link from 'next/link'
 import { ChevronLeft } from 'lucide-react'
