+++ src/app/docs/components/button/stitches-button/page.mdx
@@ -1,7 +1,7 @@
 import PreviewCodeTabs from '@/components/CodePreview/page'
 import Link from 'next/link'
 import { ChevronLeft } from 'lucide-react'
-import StitchesButton from '@/components/CodePreview/stitches-button'
+import StitchesButton from '@/showcase/components/button/StitchesButton'
 
 export const metadata = {
   title: 'Stitches Button',
