+++ src/app/docs/components/button/three-d-button/page.mdx
@@ -1,5 +1,5 @@
 import PreviewCodeTabs from '@/components/CodePreview/page'
-import { ThreeDButton } from '@/ui/button'
+import { ThreeDButton } from '@/showcase/components/button/ThreeDButton'
 import Link from 'next/link'
 import { ChevronLeft } from 'lucide-react'
