+++ src/app/docs/page.mdx
@@ -1,6 +1,6 @@
 import { RequiresTailwind } from '@/components/Requires'
 import ComponentCards from '@/components/ComponentCards'
-import AnimationCards from '@/components/AnimationCards'
+import AnimationCards from '@/components/cards/AnimationCards'
 import EffectCards from '@/components/EffectCards'
 import HooksCards from '@/components/HooksCards'
 import Image from 'next/image'
