+++ src/components/Cards/ButtonCards.tsx
@@ -1,8 +1,8 @@
 import Link from 'next/link'
-import { ThreeDButton } from '@/ui/button'
-import HeartbeatButton from '@/ui/animation/HeartbeatButton'
-import NeubrutalismButton from '../../showcase/components/button/NeubrutalismButton'
-import StitchesButton from '@/components/CodePreview/stitches-button'
+import { ThreeDButton } from '@/showcase/components/button/ThreeDButton'
+import HeartbeatButton from '@/showcase/components/button/HeartbeatButton'
+import NeubrutalismButton from '@/showcase/components/button/NeubrutalismButton'
+import StitchesButton from '@/showcase/components/button/StitchesButton'
 
 const data = [
   {
